# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories": {
      "get": {
        "description": "Get all categories",
        "operationId": "CategoryController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all categories",
        "tags": [
          "/categories"
        ]
      },
      "post": {
        "description": "Create a new category",
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create category",
        "tags": [
          "/categories"
        ]
      }
    },
...
    "/categories/slug/{slug}": {
      "get": {
        "description": "Get a single category (with subcategories) by slug",
        "operationId": "CategoryController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get category by slug",
        "tags": [
          "/categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "description": "Get a single category by id",
        "operationId": "CategoryController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get category by id",
        "tags": [
          "/categories"
        ]
      },
      "patch": {
        "description": "Update an existing category",
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update category",
        "tags": [
          "/categories"
        ]
      },
      "put": {
        "description": "Fully replace a category (all fields required)",
        "operationId": "CategoryController_replace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Replace category",
        "tags": [
          "/categories"
        ]
      },
      "delete": {
        "description": "Delete a category",
        "operationId": "CategoryController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete category",
        "tags": [
          "/categories"
        ]
      }
    },
    "/categories/{id}/subcategories": {
      "get": {
        "description": "Get all subcategories of a category",
        "operationId": "CategoryController_findSubcategories",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get subcategories",
        "tags": [
          "/categories"
        ]
      },
      "post": {
        "description": "Add a subcategory to a category",
        "operationId": "CategoryController_addSubcategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubcategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add subcategory",
        "tags": [
          "/categories"
        ]
      }
    },
    "/categories/{id}/subcategories/{subId}": {
      "get": {
        "description": "Get a specific subcategory by id",
        "operationId": "CategoryController_findSubcategoryById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get subcategory by id",
        "tags": [
          "/categories"
        ]
      },
      "patch": {
        "description": "Update an embedded subcategory",
        "operationId": "CategoryController_updateSubcategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubcategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update subcategory",
        "tags": [
          "/categories"
        ]
      },
      "put": {
        "description": "Fully replace a subcategory (all fields required)",
        "operationId": "CategoryController_replaceSubcategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubcategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Replace subcategory",
        "tags": [
          "/categories"
        ]
      },
      "delete": {
        "description": "Remove an embedded subcategory",
        "operationId": "CategoryController_removeSubcategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove subcategory",
        "tags": [
          "/categories"
        ]
      }
    },
...
    "/cart/merge": {
      "post": {
        "description": "Called after login to sync the client-side cart. If the server cart is empty, client items are validated against stock and set. If the server cart already has items, it is returned as-is and the client-side cart is discarded.",
        "operationId": "CartController_mergeCart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Merge cart",
        "tags": [
          "/cart"
        ]
      }
    },
    "/cart/items": {
      "post": {
        "description": "Add a variant to the cart. If the variant is already present, its quantity is incremented by the given amount. Stock is validated.",
        "operationId": "CartController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add item to cart",
        "tags": [
          "/cart"
        ]
      }
    },
    "/cart/items/{variantId}": {
      "patch": {
        "description": "Set the absolute quantity for a specific cart item. Stock is validated.",
        "operationId": "CartController_updateItem",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update cart item quantity",
        "tags": [
          "/cart"
        ]
      },
      "delete": {
        "description": "Remove a specific variant from the cart.",
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove item from cart",
        "tags": [
          "/cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "Urban Tab API",
    "description": "API документація",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "example": "my-slug",
            "description": "URL-friendly identifier"
          },
          "image": {
            "type": "string",
            "example": "https://cdn.example.com/categories/abc123/uuid.webp",
            "description": "Category image URL — either a direct link or the publicUrl returned from the S3 upload flow"
          },
          "order": {
            "type": "number",
            "example": 1,
            "description": "Display order of the category in the UI — lower values appear first"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "example": "my-slug",
            "description": "URL-friendly identifier"
          },
          "image": {
            "type": "string",
            "example": "https://cdn.example.com/categories/abc123/uuid.webp",
            "description": "Category image URL — either a direct link or the publicUrl returned from the S3 upload flow"
          },
          "order": {
            "type": "number",
            "example": 1,
            "description": "Display order of the category in the UI — lower values appear first"
          }
        }
      },
      "RequiredAttributeDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Виробник",
            "description": "Attribute label (human-readable)"
          },
          "filter_type": {
            "type": "string",
            "example": "multi-select",
            "enum": [
              "multi-select",
              "range"
            ],
            "description": "Filter UI type"
          },
          "unit": {
            "type": "object",
            "example": "мм",
            "description": "Unit of measure",
            "nullable": true
          }
        },
        "required": [
          "label",
          "filter_type",
          "unit"
        ]
      },
      "CreateSubcategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Smartphones",
            "description": "Subcategory name"
          },
          "slug": {
            "type": "string",
            "example": "my-slug",
            "description": "URL-friendly identifier"
          },
          "required_attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequiredAttributeDto"
            }
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
...
      "MergeCartItemDto": {
        "type": "object",
        "properties": {
          "variant_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011",
            "description": "MongoDB ObjectId of the product variant"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Number of units (minimum 1)",
            "minimum": 1
          }
        },
        "required": [
          "variant_id",
          "quantity"
        ]
      },
      "MergeCartDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MergeCartItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "variant_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011",
            "description": "MongoDB ObjectId of the product variant"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Number of units (minimum 1)",
            "minimum": 1
          }
        },
        "required": [
          "variant_id",
          "quantity"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Number of units (minimum 1)",
            "minimum": 1
          }
        },
        "required": [
          "quantity"
        ]
      },
...
    }
  }
}